+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+Revert passing BINARY 0x00 for (N)VARCHAR arguments.
 
 ## [1.6.0] - 2017-11-17
 ### Added