+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.6.2] - 2017-11-27
 ### Fixed
 - Support values longer than 8000 characters in `ctds.SqlVarChar`.
 
@@ -172,7 +174,8 @@ _NVARCHAR_ arguments to remote procedure calls.
 ## [1.0.0] - 2016-03-14
 Initial Release
 
-[Unreleased]: https://github.com/zillow/ctds/compare/v1.6.1...HEAD
+[Unreleased]: https://github.com/zillow/ctds/compare/v1.6.2...HEAD
+[1.6.2]: https://github.com/zillow/ctds/compare/v1.6.2...v1.6.1
 [1.6.1]: https://github.com/zillow/ctds/compare/v1.6.1...v1.6.0
 [1.6.0]: https://github.com/zillow/ctds/compare/v1.6.0...v1.5.0
 [1.5.0]: https://github.com/zillow/ctds/compare/v1.5.0...v1.4.1